Adjust more places to ssize_t Size/Length/Dict_Next.
git-svn-id: http://svn.python.org/projects/python/branches/ssize_t@41905 6015fed2-1504-0410-9fe1-9d1591cc4771

+++ Objects/funcobject.c
@@ -529,7 +529,7 @@ function_call(PyObject *func, PyObject *arg, PyObject *kw)
 	}
 
 	if (kw != NULL && PyDict_Check(kw)) {
-		int pos, i;
+		Py_ssize_t pos, i;
 		nk = PyDict_Size(kw);
 		k = PyMem_NEW(PyObject *, 2*nk);
 		if (k == NULL) {
